Lint fix (#12)
Now that I'm using WebStorm I'm changing some linting options just to be consistent with the toolset. We are now always using prettier to format code and eslint to validate linting.
This has forced us to rewrite a lot of the current classes.
Besides linting things I imported momentjs to start working with dates and made the day selector buttons to change the current day.